changes: service
id validation & create() optimized

+++ src/services/users.service.js
@@ -1,6 +1,7 @@
 import userRepository from "../repositories/user.repository.js"
 import bcrypt from "bcrypt"
 import AppError from "../error/error.js"
+import mongoose from "mongoose"
 
 class UserServices {
    async getAll() {
@@ -28,6 +29,10 @@ class UserServices {
 
    async getById(id) {
       try {
+         if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new AppError("Invalid product ID format", 400)
+         }
+
          const user = await userRepository.getById(id)
 
          if (!user) {
@@ -41,29 +46,29 @@ class UserServices {
       }
    }
 
-   async create(arr) {
+   async create(user) {
       try {
-         for (const user of arr) {
-            const { first_name, last_name, email, password, age, role } = user
+         const { first_name, last_name, email, password, age, role } = user
 
-            if (
-               !first_name ||
-               !last_name ||
-               !email ||
-               !password ||
-               !age ||
-               !role
-            ) {
-               throw new AppError("Missing values", 400)
-            }
-
-            const hash = await bcrypt.hash(password, 10)
-            user.password = hash
+         if (!first_name || !last_name || !email || !password || age == null) {
+            throw new AppError("Missing values", 400)
          }
 
-         const result = await userRepository.create(arr)
-         return result
+         const hashedPassword = await bcrypt.hash(password, 10)
+
+         const userToCreate = {
+            first_name,
+            last_name,
+            email,
+            age: Number(age),
+            role: role ?? "user",
+            password: hashedPassword
+         }
+
+         const userCreate = await userRepository.create(userToCreate)
+         return userCreate
       } catch (error) {
+         console.log(error)
          if (error instanceof AppError) throw error
          throw new AppError("Database error", 500)
       }
@@ -71,6 +76,10 @@ class UserServices {
 
    async update(id, data) {
       try {
+         if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new AppError("Invalid product ID format", 400)
+         }
+
          if (!data || Object.keys(data).length === 0) {
             throw new AppError("No data to update", 400)
          }
@@ -99,6 +108,10 @@ class UserServices {
 
    async delete(id) {
       try {
+         if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new AppError("Invalid product ID format", 400)
+         }
+
          const user = await userRepository.delete(id)
 
          if (!user) {
@@ -137,7 +150,6 @@ class UserServices {
    // reset password
    async resetPassword(hashedToken, newPassword) {
       try {
-         console.log(4)
          const user = await userRepository.getByResetToken(hashedToken)
 
          if (!user || user.resetPasswordExpires < Date.now()) {
